requirements.txt
amqp==5.1.1
arabic-reshaper==3.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
asn1crypto==1.5.1
astroid==2.15.5
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
bcrypt==4.0.1
billiard==4.1.0
boto3==1.26.151
botocore==1.29.151
celery==5.3.0
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.1
cssselect2==0.7.0
daphne==4.0.0
defusedxml==0.7.1
dill==0.3.6
Django==4.2.2
django-allauth==0.54.0
django-bower==5.2.0
django-ckeditor==6.5.1
django-cors-headers==4.0.0
django-crispy-forms==2.0
django-filter==23.3
django-js-asset==2.0.0
django-mathfilters==1.0.0
django-redis==5.2.0
django-resized==1.0.2
django-rest-auth==0.9.5
django-storages==1.13.2
django-taggit==4.0.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
dnspython==2.3.0
environ==1.0
et-xmlfile==1.1.0
html5lib==1.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
kombu==5.3.0
lazy-object-proxy==1.9.0
lxml==4.9.3
Markdown==3.4.3
MarkupSafe==2.1.3
mccabe==0.7.0
msgpack==1.0.6
mutagen==1.46.0
numpy==1.26.1
oauthlib==3.2.2
openpyxl==3.1.2
oscrypto==1.3.0
pandas==2.1.1
Pillow==9.5.0
platformdirs==3.5.3
prompt-toolkit==3.0.38
psycopg==3.1.10
psycopg-binary==3.1.10
psycopg2==2.9.6
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyHanko==0.20.1
pyhanko-certvalidator==0.24.1
PyJWT==2.7.0
pylint==2.17.4
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
pymongo==4.3.3
pyOpenSSL==23.2.0
pypdf==3.16.4
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
qrcode==7.4.2
redis==4.5.5
reportlab==3.6.13
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.6.1
service-identity==23.1.0
six==1.16.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.11.8
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.2
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
whitenoise==6.4.0
wrapt==1.15.0
xhtml2pdf==0.2.11
xlrd==2.0.1
xlwt==1.3.0
zope.interface==6.0
